Add logging Add controllers Update model Refactor services Add decoder

+++ src/main/java/com/stdevi/dragonsofmugloar/model/game/Reputation.java
@@ -6,7 +6,7 @@ import lombok.ToString;
 @Data
 @ToString
 public class Reputation {
-    private long people;
-    private long state;
-    private long underworld;
+    private double people;
+    private double state;
+    private double underworld;
 }
